Rename AppControl to ClientControl
