+++ Assets/Holiday/Controls/MultiplayControl/MultiplayControl.unity
@@ -369,6 +369,6 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   parentReference:
-    TypeName: Extreal.SampleApp.Holiday.Controls.AppControl.AppControlScope
+    TypeName: Extreal.SampleApp.Holiday.Controls.ClientControl.ClientControlScope
   autoRun: 1
   autoInjectGameObjects: []
