+++ Assets/Holiday/Controls/TextChatControl/TextChatControl.unity
@@ -153,7 +153,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   parentReference:
-    TypeName: Extreal.SampleApp.Holiday.Controls.AppControl.AppControlScope
+    TypeName: Extreal.SampleApp.Holiday.Controls.ClientControl.ClientControlScope
   autoRun: 1
   autoInjectGameObjects: []
   textChatControlView: {fileID: 685982609}
