+++ Assets/Holiday/Screens/AvatarSelectionScreen/AvatarSelectionScreen.unity
@@ -2259,7 +2259,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   parentReference:
-    TypeName: Extreal.SampleApp.Holiday.Controls.AppControl.AppControlScope
+    TypeName: Extreal.SampleApp.Holiday.App.AppScope
   autoRun: 1
   autoInjectGameObjects: []
   avatarSelectionScreenView: {fileID: 1514294814}
